feat: 新增SQL效验
