+++ gear4j-safe/src/main/java/club/gclmit/gear4j/safe/config/Gear4jSafeProperties.java
@@ -138,125 +138,61 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package club.gclmit.gear4j.safe.xss;
+package club.gclmit.gear4j.safe.config;
 
-import org.jsoup.Jsoup;
-import org.jsoup.internal.StringUtil;
-import org.jsoup.nodes.Attribute;
-import org.jsoup.nodes.Document;
-import org.jsoup.nodes.Element;
-import org.jsoup.nodes.Entities;
-import org.springframework.web.util.HtmlUtils;
+import java.util.ArrayList;
+import java.util.List;
 
-import club.gclmit.gear4j.core.lang.Log;
-import club.gclmit.gear4j.core.lang.LoggerProvider;
-import club.gclmit.gear4j.core.utils.StringUtils;
-import club.gclmit.gear4j.safe.config.Gear4jXssProperties;
-import cn.hutool.core.util.CharsetUtil;
+import org.springframework.boot.context.properties.ConfigurationProperties;
 
 /**
- * Xss Utils
+ * Xss配置类
  *
  * @author <a href="https://blog.gclmit.club">gclm</a>
  */
-public class XssHandler {
+@ConfigurationProperties(prefix = Gear4jSafeProperties.PREFIX)
+public class Gear4jSafeProperties {
 
-    private XssHandler() {}
+    public static final String PREFIX = "gear4j.safe";
 
-    public static final HtmlSafeList WHITE_LIST = HtmlSafeList.INSTANCE;
-
-    private static final Document.OutputSettings DEFAULT_OUTPUT_SETTINGS =
-        new Document.OutputSettings().prettyPrint(false);
+    public static final String CONFIG_NAME = "gear4jSafeProperties";
 
     /**
-     * 基于properties 处理xss <br>
-     * 1. 非空判断 2. 判断处理模式，如果是转义，直接调用spring工具类处理 3. 如果是clean模式，需要根据自定义情况去配置清理参数
-     *
-     * @param properties xss 配置参数
-     * @param text text
-     * @return {@link String}
+     * 开启安全配置（SQL、xss）
      */
-    public static String clean(Gear4jXssProperties properties, String text) {
-        if (StringUtil.isBlank(text)) {
-            return text;
-        }
-        String result = text;
-        Gear4jXssProperties.Mode mode = properties.getMode();
-        if (Gear4jXssProperties.Mode.escape == mode) {
-            result = HtmlUtils.htmlEscape(text, CharsetUtil.UTF_8);
-        } else {
-            Document.OutputSettings outputSettings = new Document.OutputSettings()
-                // 这个规则最少
-                .escapeMode(Entities.EscapeMode.xhtml).prettyPrint(properties.isPrettyPrint());
-            String escapedHtml = Jsoup.clean(text, "", WHITE_LIST, outputSettings);
+    private boolean enabled = false;
 
-            // 转义判断，目前clean默认转义
-            result = properties.isEnableEscape() ? escapedHtml : Entities.unescape(escapedHtml);
-        }
+    /**
+     * 拦截的路由，默认为/*
+     */
+    private List<String> urlPatterns = List.of("/*");
 
-        if (properties.isLog()) {
-            Log.debug(LoggerProvider.GEAR4J, "Status: {}\tValue:[{}] -> [{}]", !result.equals(text), text, result);
-        }
-        return result;
+    /**
+     * 放行的路由，默认为空
+     */
+    private List<String> excludes = new ArrayList<>();
+
+    public boolean isEnabled() {
+        return enabled;
     }
 
-    /**
-     * 自定义白名单
-     */
-    public static class HtmlSafeList extends org.jsoup.safety.Safelist {
+    public void setEnabled(boolean enabled) {
+        this.enabled = enabled;
+    }
 
-        public static final HtmlSafeList INSTANCE = new HtmlSafeList();
+    public List<String> getUrlPatterns() {
+        return urlPatterns;
+    }
 
-        public HtmlSafeList() {
-            addTags("a", "b", "blockquote", "br", "caption", "cite", "code", "col", "colgroup", "dd", "div", "span",
-                "embed", "object", "dl", "dt", "em", "h1", "h2", "h3", "h4", "h5", "h6", "i", "img", "li", "ol", "p",
-                "pre", "q", "small", "strike", "strong", "sub", "sup", "table", "tbody", "td", "tfoot", "th", "thead",
-                "tr", "u", "ul");
+    public void setUrlPatterns(List<String> urlPatterns) {
+        this.urlPatterns = urlPatterns;
+    }
 
-            addAttributes("a", "href", "title", "target");
-            addAttributes("blockquote", "cite");
-            addAttributes("col", "span");
-            addAttributes("colgroup", "span");
-            addAttributes("img", "align", "alt", "src", "title");
-            addAttributes("ol", "start");
-            addAttributes("q", "cite");
-            addAttributes("table", "summary");
-            addAttributes("td", "abbr", "axis", "colspan", "rowspan", "width");
-            addAttributes("th", "abbr", "axis", "colspan", "rowspan", "scope", "width");
-            addAttributes("video", "src", "autoplay", "controls", "loop", "muted", "poster", "preload");
-            addAttributes("object", "width", "height", "classid", "codebase");
-            addAttributes("param", "name", "value");
-            addAttributes("embed", "src", "quality", "width", "height", "allowFullScreen", "allowScriptAccess",
-                "flashvars", "name", "type", "pluginspage");
+    public List<String> getExcludes() {
+        return excludes;
+    }
 
-            addAttributes(":all", "class", "style", "height", "width", "type", "id", "name");
-
-            addProtocols("blockquote", "cite", "http", "https");
-            addProtocols("cite", "cite", "http", "https");
-            addProtocols("q", "cite", "http", "https");
-
-            // 如果添加以下的协议，那么href 必须是http、 https 等开头，相对路径则被过滤掉了
-            // addProtocols("a", "href", "ftp", "http", "https", "mailto", "tel");
-
-            // 如果添加以下的协议，那么src必须是http 或者 https 开头，相对路径则被过滤掉了，
-            // 所以必须注释掉，允许相对路径的图片资源
-            // addProtocols("img", "src", "http", "https");
-        }
-
-        @Override
-        protected boolean isSafeAttribute(String tagName, Element el, Attribute attr) {
-            // 不允许 javascript 开头的 src 和 href
-            if ("src".equalsIgnoreCase(attr.getKey()) || "href".equalsIgnoreCase(attr.getKey())) {
-                String value = attr.getValue();
-                if (StringUtils.isNotBlank(value) && value.toLowerCase().startsWith("javascript")) {
-                    return false;
-                }
-            }
-            // 允许 base64 的图片内容
-            if ("img".equals(tagName) && "src".equals(attr.getKey()) && attr.getValue().startsWith("data:;base64")) {
-                return true;
-            }
-            return super.isSafeAttribute(tagName, el, attr);
-        }
+    public void setExcludes(List<String> excludes) {
+        this.excludes = excludes;
     }
 }
