+++ gear4j-safe/src/main/java/club/gclmit/gear4j/safe/core/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * safe核心处理类
+ *
+ * @author <a href="https://blog.gclmit.club">gclm</a>
+ * @since 2022/7/5 15:40
+ * @since jdk11
+ */
+package club.gclmit.gear4j.safe.core;
