+++ gear4j-safe/src/main/java/club/gclmit/gear4j/safe/handler/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * 效验处理器
+ *
+ * @author <a href="https://blog.gclmit.club">gclm</a>
+ * @since 2022/7/5 15:40
+ * @since jdk11
+ */
+package club.gclmit.gear4j.safe.handler;
