--- gear4j-safe/src/main/java/club/gclmit/gear4j/safe/xss/package-info.java
@@ -1,8 +1,0 @@
-/**
- * gear4j-waf: waf防御工具类
- *
- * @author <a href="https://blog.gclmit.club">gclm</a>
- * @since 2022/1/10 2:35 PM
- * @since jdk11
- */
-package club.gclmit.gear4j.safe.xss;
