+++ gear4j-safe/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,2 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-club.gclmit.gear4j.safe.config.Gear4jXssConfiguration
+club.gclmit.gear4j.safe.config.Gear4jSafeConfiguration
